start xata switch, working on inactivity checks

See merge request glass-ships/compass-bot!9

+++ src/compass_bot/cogs/listeners.py
@@ -60,43 +60,54 @@ class Listeners(commands.Cog):
         await ctx.send(embed=discord.Embed(description=message))  # , delete_after=60)
         # await ctx.message.delete(delay=5)
 
-    @commands.Cog.listener()
-    async def on_message(self, message):
-        """Move videos"""
-
+    @commands.Cog.listener("on_message")
+    async def move_videos(self, message: discord.Message):
+        """Move videos to the specified video channel, if it is set."""
         if message.author.bot:
             return
-
         vid_channel = bot.db.get_channel_vids(message.guild.id)
-        if vid_channel == 0 or message.channel.id != vid_channel:
+        if vid_channel is None or message.channel.id == vid_channel:
             return
-
         vid_links = ["youtube.com/watch?", "youtu.be/", "vimeo.com/", "dailymotion.com/video", "tiktok.com"]
-        newmessage = f"{message.author.name} has uploaded a video.\n─── **Original Message** ───\n\n{message.content}\n"
-
-        if any(i in message.content for i in vid_links):
+        if not any(i in message.content for i in vid_links) and not any(
+            "video" in a.content_type for a in message.attachments
+        ):
+            return
+        else:
             files = []
             if message.attachments:
-                for a in filter(lambda x: x.size < message.guild.filesize_limit, message.attachments):
+                for a in filter(
+                    # lambda x,: x.size < message.guild.filesize_limit and "video" in x.content_type, message.attachments
+                    lambda x,: x.size < message.guild.filesize_limit,
+                    message.attachments,
+                ):
                     await download(message, a, "temp/moved_messages")
                     files.append(discord.File(getfilepath(message, f"temp/moved_messages/{a.filename}")))
+            newmessage = (
+                f"{message.author.mention} has uploaded a video.\n─── **Original Message** ───\n\n{message.content}\n"
+            )
             if any(a.size >= message.guild.filesize_limit for a in message.attachments):
                 newmessage += "`Plus some files too large to resend`"
-
             # Move the message
             chan = await bot.fetch_channel(vid_channel)
             movedmessage = await chan.send(content=newmessage, files=files)
-
             # Delete original and notify author
             await message.delete()
             await message.channel.send(
                 f"""
 {message.author.mention} - your message contained a video, and was moved to <#{vid_channel}>.
-**Link:** \nhttps://discord.com/channels/{message.guild.id}/{vid_channel}/{movedmessage.id}
-            """
+**Link:** https://discord.com/channels/{message.guild.id}/{vid_channel}/{movedmessage.id}
+"""
             )
-
             return
+
+    @commands.Cog.listener("on_message")
+    async def log_activity(self, message: discord.Message):
+        """Log messages to the database."""
+        if message.author.bot:
+            return
+        if message.guild:
+            bot.db.add_or_update_user_log(message.guild.id, message.author.id, message.created_at)
 
     @commands.Cog.listener()
     async def on_raw_reaction_add(self, payload):  # reaction, user):
@@ -108,13 +119,18 @@ class Listeners(commands.Cog):
         reaction = payload.emoji
         emoji = "🔍"
         user = payload.member
-
+        channel_id = bot.db.get_channel_logs(message.guild.id)
+        log_channel = get(message.guild.text_channels, id=channel_id)
+        if not log_channel:
+            # await message.channel.send(
+            #     f"Logs channel has not been set!\nUse `;set_logs_channel <#channel>` to set one."
+            # )
+            return
         if reaction.name == emoji:
             # Check for mod role
-
             mod_roles = bot.db.get_mod_roles(payload.guild_id)
             roles = [x.id for x in user.roles]
-            if any(r in roles for r in mod_roles):
+            if any(int(r) in roles for r in mod_roles):
                 # Remove the reaction
                 await message.remove_reaction("🔍", user)
 
@@ -141,14 +157,8 @@ class Listeners(commands.Cog):
                 ce.add_field(name="Flagged at:", value=f"<t:{flagged_time}:R>")
 
                 # Send embed to logs
-                channel_id = bot.db.get_channel_logs(flagged_message.guild.id)
-                channel = get(flagged_message.guild.text_channels, id=channel_id)
-                if channel:
-                    await channel.send(content=None, embed=ce)
-                else:
-                    await flagged_message.channel.send(
-                        f"Logs channel has not been set!\nUse `;set_logs_channel <#channel>` to set one."
-                    )
+                await log_channel.send(content=None, embed=ce)
+
         return
 
     #######################
@@ -163,9 +173,9 @@ class Listeners(commands.Cog):
         channel = get(glass_guild.text_channels, id=chan_id)
 
         data = GuildData(guild).__dict__
-        del data["Guild"]
+        del data["guild"]
 
-        if bot.db.add_guild_table(guild.id, data):
+        if bot.db.add_guild(guild.id, data):
             await channel.send(embed=discord.Embed(description=f'Guild "{guild.name}" added to database.'))
         else:
             await channel.send(embed=discord.Embed(description=f'Guild "{guild.name}" already in database.'))
@@ -200,6 +210,9 @@ class Listeners(commands.Cog):
         import httpx
 
         guild = member.guild
+        if guild.id != GLASS_HARBOR:
+            return
+
         channel_id = bot.db.get_channel_welcome(guild_id=guild.id)
         if channel_id == 0:
             return
@@ -209,14 +222,11 @@ class Listeners(commands.Cog):
         ocean = Image.open(f"{COMPASS_SRC}/images/welcome_background.png")
         pfp = Image.open(io.BytesIO(av.content))
         pfp = pfp.resize((650, 650))
-
         ocean_w, ocean_h = ocean.size
         pfp_w, pfp_h = pfp.size
         offset = ((ocean_w - pfp_w) // 2, (ocean_h - pfp_h) // 2)
-
         base = Image.new("RGBA", ocean.size)
         base.paste(pfp, offset)
-
         img = Image.alpha_composite(base, ocean)
 
         with io.BytesIO() as image_binary:
@@ -236,3 +246,15 @@ class Listeners(commands.Cog):
                 ),
                 file=discord.File(fp=image_binary, filename="image.png"),
             )
+
+    @commands.Cog.listener()
+    async def on_member_remove(self, member: discord.Member) -> None:
+        guild = member.guild
+        # channel_id = bot.db.get_channel_welcome(guild_id=guild.id)
+        # if channel_id is None:
+        #     return
+        # channel = get(guild.text_channels, id=channel_id)
+        # await channel.send(
+        #     embed=discord.Embed(title=f"Goodbye, {member.name}!", description=f"{member.mention} has left the server.")
+        # )
+        bot.db.remove_user_log(guild.id, member.id)
